% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.oe_cache_dir}
\alias{.oe_cache_dir}
\title{Package cache directory}
\usage{
.oe_cache_dir()
}
\value{
A path (character).
}
\description{
Returns (and creates) the per-user cache directory used by \code{oneearthr}
for downloaded layers.
}
\keyword{internal}
