% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oe_data}
\alias{oe_data}
\alias{oe_realms}
\alias{oe_subrealms}
\alias{oe_bioregions}
\alias{oe_ecoregions}
\title{One Earth Bioregions Framework layers (packaged)}
\format{
Wrapped \code{SpatVector} polygon layers in EPSG:4326, simplified to
keep the package lightweight (use \code{oe_layer(layer, full = TRUE)} for the
original geometries):
\describe{
\item{oe_realms}{14 biogeographical realms (~0.02 deg tolerance).}
\item{oe_subrealms}{53 subrealms (~0.02 deg).}
\item{oe_bioregions}{185 bioregions, coded by realm prefix + number
(~0.02 deg).}
\item{oe_ecoregions}{847 ecoregions from RESOLVE Ecoregions 2017
(~0.15 deg).}
}
}
\source{
One Earth Bioregions Framework (2023), \url{https://www.oneearth.org},
licensed CC BY-NC 4.0 (non-commercial use, cite One Earth);
Ecoregions: Dinerstein et al. (2017) \doi{10.1093/biosci/bix014},
licensed CC-BY 4.0.
}
\usage{
oe_realms

oe_subrealms

oe_bioregions

oe_ecoregions
}
\description{
Packaged spatial layers of the One Earth Bioregions Framework, stored as
wrapped \code{terra::SpatVector} objects (see \link[terra:wrap]{terra::wrap}). Use \code{\link[=oe_layer]{oe_layer()}}
or \code{\link[terra:unwrap]{terra::unwrap()}} to materialise them.
}
\keyword{datasets}
