% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe_layer.R
\name{oe_layers}
\alias{oe_layers}
\title{List the available One Earth layers}
\usage{
oe_layers()
}
\value{
A \code{data.frame} describing the four framework layers.
}
\description{
List the available One Earth layers
}
\examples{
oe_layers()
}
