% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_exists}
\alias{url_exists}
\title{Check that a URL is reachable}
\usage{
url_exists(x, quiet = FALSE, ...)
}
\arguments{
\item{x}{A string with the url to check.}

\item{quiet}{Logical; suppress the error message.}

\item{...}{Passed to \link[utils:download.file]{utils::download.file} internals.}
}
\value{
Logical, \code{TRUE} if the url responds.
}
\description{
Check that a URL is reachable
}
\keyword{internal}
